On déplace le détail des aides dans le formulaire principale car il se peut que rule soit vide et que InputSwitch ne soit pas déclenché --> page blanche
