+++ components/ampleur/CEEAmpleur.tsx
@@ -3,7 +3,6 @@ import AideAmpleur from './AideAmpleur'
 
 export default function CEEAmpleur({
   isEligible,
-  rules,
   engine,
   situation,
   setSearchParams,
@@ -11,10 +10,6 @@ export default function CEEAmpleur({
   expanded,
 }) {
   const dottedName = "CEE . rénovation d'ampleur"
-  const value = situation['projet . DPE visé'],
-    oldIndex = +situation['DPE . actuel'] - 1,
-    automaticChoice = Math.max(oldIndex - 2, 0),
-    choice = value ? Math.min(automaticChoice, value - 1) : automaticChoice
 
   return (
     <AideAmpleur
@@ -36,9 +31,6 @@ export default function CEEAmpleur({
           </p>
           <CEEAmpleurScenario
             {...{
-              rules,
-              choice,
-              oldIndex,
               engine,
               situation,
               setSearchParams,
